backend/requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
Flask-Cors==4.0.0
pymongo==4.6.1
pycryptodome==3.20.0
bcrypt==4.1.2
python-dotenv==1.0.0
email-validator==2.1.0
